chore: remove all tracking
