+++ src/components/navigate.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { useAnalytics } from '@common/hooks/use-analytics';
+import { useDoChangeScreen } from '@common/hooks/use-do-change-screen';
 import { ScreenPaths } from '@store/onboarding/types';
 import { useNavigate, ToOptions } from 'react-router-dom';
 
@@ -9,7 +9,7 @@ interface NavigateProps {
 }
 
 export const Navigate: React.FC<NavigateProps> = ({ to, screenPath }) => {
-  const { doChangeScreen } = useAnalytics();
+  const doChangeScreen = useDoChangeScreen();
   const navigate = useNavigate();
   useEffect(() => {
     navigate(to);
